Fix build on Windows (mingw)
Added a few small changes to the main makefile to allow building the emulator directly on Windows using MinGW and GNU Make.

Also added CoreUtils for Windows as build prerequisite; see README.md for instructions how to install it. This package provides Windows versions of some Linux commands like cp, test, and others, which minimizes the need for OS-specific logic in the makefile.