Fix a build error on Windows
The recent introduction of system-deps broke the build on Windows, apparently because system-deps uses pkg-config internally, and Windows does not have it. The quick and dirty fix is make system-deps to not run on Windows targets (Linux and MacOS supposedly have pkg-config).

Also bumped to a newer version of system-deps.